feat: Add a comprehensive file downloader and scraper with logging and reporting
- Implemented a parallel file downloading mechanism using Ray for improved performance.
- Added detailed logging functionality with color-coded messages for different log levels.
- Introduced a state management system to track completed and failed downloads.
- Created a report generation feature that includes statistics and visualizations of download delays.
- Integrated BeautifulSoup for HTML parsing to extract file links based on specified patterns.
- Enhanced error handling and retry logic for robust downloading.
- Configured customizable headers and user agents for requests.
- Established a command-line interface (CLI) for user interaction and configuration.
